+++ NEWS.md
@@ -1,8 +1,12 @@
-# recsub Release Notes
+# `oshka` Release Notes
 
-## v0.0.2
+## v0.2.0
 
-* [#5](https://github.com/brodieG/recsub/issues/5): Substitution now respects
+* [#1](https://github.com/brodieG/oshka/issues/1): Provide mechanism for
+  shielding some elements from expansion.
+* [#2](https://github.com/brodieG/oshka/issues/2): Function symbols
+  substituted.
+* [#5](https://github.com/brodieG/oshka/issues/5): Substitution now respects
   search path order.
 
 ## v0.0.1